frontend: rename files
