+++ frontend/src/components/Main.tsx
@@ -1,8 +1,8 @@
 import React, { Fragment, useContext } from "react";
 import { Navigate } from "react-router-dom";
 import AuthContext from "../contexts/auth-context";
-import NewTask from "./NewTask";
-import TaskList from "./TaskList";
+import NewTask from "./task/NewTask";
+import TaskList from "./task/TaskList";
 
 const Main: React.FC = () => {
   const authCtx = useContext(AuthContext);
